Fix merged photo file "Frames" int64 overflow
Who even keeps videos with more than 2 billion frames in their gallery?

+++ app/src/main/java/ua/com/radiokot/photoprism/api/photos/model/PhotoPrismMergedPhoto.kt
@@ -80,7 +80,7 @@ constructor(
         @JsonProperty("Duration")
         val duration: Long?,
         @JsonProperty("Frames")
-        val frames: Int?,
+        val frames: Long?,
         @JsonProperty("Primary")
         val primary: Boolean?,
         @JsonProperty("Root")
